changed displaySize reducer name to generalUISettings | moved mock data location
