+++ src/components/header/Header.tsx
@@ -4,20 +4,20 @@ import MenuIcon from "../../assets/images/icons/menu-logo.svg";
 import DishesIcon from "../../assets/images/icons/dishes-icon.png";
 import NavLinks from "../nav-links/NavLinks";
 import Search from "../search/Search";
-import { mockNavMainItems, mockNavUserItems } from "./mockNavData";
 import { useAppSelector } from "../../state/hooks";
 import { RootState } from "../../state/store";
+import { mockNavMainItems, mockNavUserItems } from "../../utils/constants/mockNavData";
 
 const Header: React.FC = () => {
-  const displaySizeState = useAppSelector(
-    (state: RootState) => state.displaySize.displaySize
+  const generalUISettingsState = useAppSelector(
+    (state: RootState) => state.generalUISettings.displaySize
   );
 
   return (
     <header className="header-container flex row space-between align-center w-all">
       <nav className="nav-main flex space-between align-center">
         <div className="main-links">
-          {displaySizeState === "mobile" ? (
+          {generalUISettingsState === "mobile" ? (
             <img src={MenuIcon} alt="menu-icon" />
           ) : (
             <NavLinks navItems={mockNavMainItems} styleClassNames={"flex row flex space-around align-center w-all"} />
