+++ src/utils/constants/mockNavData.ts
@@ -1,4 +1,4 @@
-import navItem from "../../utils/interfaces/navItem";
+import navItem from "../interfaces/navItem";
 import UserIcon from "../../assets/images/icons/user-icon.svg";
 import BagIcon from "../../assets/images/icons/bag-icon.svg";
