--- src/utils/interfaces/displaySize.ts
@@ -1,4 +1,0 @@
-type displaySize = {
-    displaySize: 'mobile' | 'tablet' | 'web'
-};
-export default displaySize;
