+++ src/utils/interfaces/generalUISettings.ts
@@ -1,0 +1,4 @@
+type GeneralUISettings = {
+    displaySize: 'mobile' | 'tablet' | 'web'
+};
+export default GeneralUISettings;
